Moves flattenObject function to utils.api.ts
This commit refactors the code by moving the `flattenObject`
function from `SpanDetailsPanel.tsx` to `utils.api.ts`. This promotes
code reuse and avoids duplication, as this utility function
might be useful in other parts of the application as well.
