+++ src/utils/utils.api.ts
@@ -100,3 +100,68 @@ export async function searchTags(datasourceUid: string, query: string, start: nu
   const response = await lastValueFrom(responses);
   return response.data.scopes.find((scope) => scope.name === 'span')?.tags || [];
 }
+
+export const flattenObject = (obj: any, prefix = ''): Record<string, any> => {
+  const flattened: Record<string, any> = {};
+
+  // Handle array of key-value pairs with typed values
+  if (Array.isArray(obj)) {
+    obj.forEach((item: any) => {
+      if (item && typeof item === 'object' && 'key' in item && 'value' in item) {
+        const key = item.key;
+        const value = item.value;
+
+        // Extract the actual value based on the type
+        let actualValue: any = null;
+        // typeof null === 'object' in JS, so we need explicit null check to avoid accessing properties on null
+        if (typeof value === 'object' && value !== null) {
+          // Check for type-specific value fields
+          if ('intValue' in value) {
+            actualValue = parseInt(value.intValue, 10);
+          } else if ('stringValue' in value) {
+            actualValue = value.stringValue;
+          } else if ('boolValue' in value) {
+            actualValue = value.boolValue;
+          } else if ('doubleValue' in value) {
+            actualValue = parseFloat(value.doubleValue);
+          } else if ('bytesValue' in value) {
+            actualValue = value.bytesValue;
+          } else {
+            // Fallback to JSON string if unknown type
+            actualValue = JSON.stringify(value);
+          }
+        } else {
+          actualValue = value;
+        }
+
+        const newKey = prefix ? `${prefix}.${key}` : key;
+        flattened[newKey] = actualValue;
+      }
+    });
+  } else if (typeof obj === 'object' && obj !== null) {
+    // Handle regular nested objects
+    for (const key in obj) {
+      if (obj.hasOwnProperty(key)) {
+        const newKey = prefix ? `${prefix}.${key}` : key;
+
+        if (Array.isArray(obj[key])) {
+          // Handle arrays with bracket notation
+          obj[key].forEach((item: any, index: number) => {
+            const arrayKey = prefix ? `${prefix}[${index}]` : `[${index}]`;
+            if (typeof item === 'object' && item !== null) {
+              Object.assign(flattened, flattenObject(item, arrayKey));
+            } else {
+              flattened[arrayKey] = item;
+            }
+          });
+        } else if (typeof obj[key] === 'object' && obj[key] !== null) {
+          Object.assign(flattened, flattenObject(obj[key], newKey));
+        } else {
+          flattened[newKey] = obj[key];
+        }
+      }
+    }
+  }
+
+  return flattened;
+};
